5/28 GameProject
- 일부 변수 형식 int로 변경
- 데미지 텍스트 띄우는 기능 추가
- 투사체 스크립트 direction 받아오는 방식 변경
- Hp바 추가
---- 버그 사항 ----
- Hp바 버그
- 원거리 몬스터 위치 못받아오는 버그

+++ GameProject/Assets/Script/Data/MonsterData.cs
@@ -9,10 +9,10 @@ namespace MonsterData
         public int monsterID;
         public string monsterName;
         public string monsterType;
-        public float monsterCurrentHp;
-        public float monsterMaxHp;
-        public float monsterAtk;
-        public float monsterDef;
+        public int monsterCurrentHp;
+        public int monsterMaxHp;
+        public int monsterAtk;
+        public int monsterDef;
         public float monsterCritChance;
         public float monsterCritDmg;
         public float monsterDefencePenetration;
@@ -25,9 +25,9 @@ namespace MonsterData
             monsterID = int.Parse(data[0]);
             monsterName = data[1];
             monsterType = data[2];
-            monsterMaxHp = float.Parse(data[3]);
-            monsterAtk = float.Parse(data[4]);
-            monsterDef = float.Parse(data[5]);
+            monsterMaxHp = int.Parse(data[3]);
+            monsterAtk = int.Parse(data[4]);
+            monsterDef = int.Parse(data[5]);
             monsterCritChance = float.Parse(data[6]);
             monsterCritDmg = float.Parse(data[7]);
             monsterDefencePenetration = float.Parse(data[8]);
